+++ src/app/data/models.ts
@@ -392,6 +392,7 @@ export interface RSLEvent {
   pin?: Pin; // Added by dust
   wa: boolean; // Wheelchair accessible
   lm?: boolean; // Live music
+  rr?: boolean; // Rock Star Librarian Recommended
   waNotes: string;
   distance: number;
   distanceInfo: string;